できたぜ

+++ Assets/Script/FireworksPreview/Json/HanabiDataList.cs
@@ -6,27 +6,27 @@
 public class HanabiDataList
 {
     // お子様名
-    public string   author       = string.Empty;
+    public string   author         = string.Empty;
     // 性別
-    public string   sex          = string.Empty;
+    public string   sex            = string.Empty;
     // 花火名
-    public string   name         = string.Empty;
+    public string   name           = string.Empty;
     // Y座標
-    public long     y            = 0;
+    public double     y            = 0;
     // スピード
-    public long     speed_main   = 0;
+    public double     speed_main   = 0;
     // 花火タイプ
-    public string   type         = string.Empty;
+    public string   type           = string.Empty;
     // スピード
-    public long     speed        = 0;
+    public double     speed        = 0;
     // 色
-    public long     color        = 0;
+    public double     color        = 0;
     // 角度
-    public long     degree       = 0;
+    public double     degree       = 0;
     // 半径
-    public long     radius       = 0;
+    public double     radius       = 0;
     // 幅
-    public long     rectWidth    = 0;
+    public double     rectWidth    = 0;
     // 高さ
-    public long     rectHeight   = 0;
+    public double     rectHeight   = 0;
 }
